% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as_tmat}
\alias{as_tmat}
\title{Coerce a panel / msm2data object to an mstate transition matrix}
\usage{
as_tmat(object, ...)
}
\arguments{
\item{object}{A "msm2data" object or a panel data frame.}

\item{...}{Passed to \code{\link{prep2}} when \code{object} is a data frame.}
}
\value{
A transition matrix as returned by \code{mstate::transMat}.
}
\description{
Coerce a panel / msm2data object to an mstate transition matrix
}
