% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep2.R
\name{from_msdata}
\alias{from_msdata}
\title{Convert an mstate 'msdata' object to discrete-time panel format}
\usage{
from_msdata(
  x,
  id = "id",
  from = "from",
  to = "to",
  Tstart = "Tstart",
  Tstop = "Tstop",
  status = "status",
  round_fun = rnd
)
}
\arguments{
\item{x}{An mstate "msdata" object (or a data frame with the named columns).}

\item{id, from, to, Tstart, Tstop, status}{Column names.}

\item{round_fun}{Discretisation function for transition times. Default rnd().}
}
\value{
A data.table in panel format (id, time, state).
}
\description{
Convert an mstate 'msdata' object to discrete-time panel format
}
