% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{rnd}
\alias{rnd}
\title{Round half up (matching the convention used in the DIVINE analysis)}
\usage{
rnd(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
\code{trunc(x + sign(x) * 0.5)}.
}
\description{
Round half up (matching the convention used in the DIVINE analysis)
}
